macro: change DECIMAL_STR_WIDTH() return type to size_t, like strlen() and so on

+++ src/basic/macro.h
@@ -350,7 +350,7 @@ static inline int __coverity_check_and_return__(int condition) {
 #define DECIMAL_STR_WIDTH(x)                            \
         ({                                              \
                 typeof(x) _x_ = (x);                    \
-                unsigned ans = 1;                       \
+                size_t ans = 1;                         \
                 while ((_x_ /= 10) != 0)                \
                         ans++;                          \
                 ans;                                    \
